Split graphql codes
