+++ webserver/Cargo.toml
@@ -10,7 +10,6 @@ version = "0.1.0"
 async-std = {version = "1.8.0", features = ["attributes"]}
 bcrypt = "0.9"
 chrono = "0.4"
-dyn-clone = "1.0"
 juniper = "0.15"
 lazy_static = "1.4.0"
 serde = {version = "1.0", features = ["derive"]}
